requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==10.4.0
python-decouple==3.8
sqlparse==0.5.1
